crypto_common

Module hazmat

Hazardous materials.

Structs

DeserializeStateError

The error type returned when an object cannot be deserialized from the state.

Traits

SerializableState

Types which can serialize the internal state and be restored from it.

Type Aliases

AddSerializedStateSize

Alias for AddSerializedStateSize<T, S> = Sum<T, S::SerializedStateSize>

SerializedState

Serialized internal state.

SubSerializedStateSize

Alias for SubSerializedStateSize<T, S> = Diff<T, S::SerializedStateSize>