Trait crypto_common::Buffer

pub trait Buffer: Sized + WriteBytesExt {
    type Result;
    fn start() -> Self;

    fn result(self) -> Self::Result;

    fn start_hint(_l: usize) -> Self { ... }
}

Trait for writers which will not fail in normal operation with small amounts of data, e.g., Vec. Moreover having a special trait allows us to implement it for other types, such as the SHA Digest.

Associated Types

type Result

Loading content...
Required methods

fn start() -> Self

fn result(self) -> Self::Result

Loading content...
Provided methods

fn start_hint(_l: usize) -> Self

Loading content...