crypto_common

Trait InnerIvInit

pub trait InnerIvInit:
    InnerUser
    + IvSizeUser
    + Sized {
    // Required method
    fn inner_iv_init(inner: Self::Inner, iv: &Iv<Self>) -> Self;

    // Provided method
    fn inner_iv_slice_init(
        inner: Self::Inner,
        iv: &[u8],
    ) -> Result<Self, InvalidLength> { ... }
}

Types which can be initialized from another type and additional initialization vector/nonce.

Usually used for initializing types from block ciphers.

Required Methods

fn inner_iv_init(inner: Self::Inner, iv: &Iv<Self>) -> Self

Initialize value using inner and iv array.

Provided Methods

fn inner_iv_slice_init( inner: Self::Inner, iv: &[u8], ) -> Result<Self, InvalidLength>

Initialize value using inner and iv slice.

Errors

Returns InvalidLength in the event that iv is not the expected length.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.