crypto_common

Trait KeyIvInit

pub trait KeyIvInit:
    KeySizeUser
    + IvSizeUser
    + Sized {
    // Required method
    fn new(key: &Key<Self>, iv: &Iv<Self>) -> Self;

    // Provided methods
    fn new_from_slices(key: &[u8], iv: &[u8]) -> Result<Self, InvalidLength> { ... }
    fn generate_key<R: CryptoRng>(rng: &mut R) -> Key<Self> { ... }
    fn generate_iv<R: CryptoRng>(rng: &mut R) -> Iv<Self> { ... }
    fn generate_key_iv<R: CryptoRng>(rng: &mut R) -> (Key<Self>, Iv<Self>) { ... }
}

Types which can be initialized from a key and initialization vector (nonce).

Required Methods

fn new(key: &Key<Self>, iv: &Iv<Self>) -> Self

Create new value from fixed length key and nonce.

Provided Methods

fn new_from_slices(key: &[u8], iv: &[u8]) -> Result<Self, InvalidLength>

Create new value from variable length key and nonce.

Errors

Returns InvalidLength in the event that key and/or iv are not the expected length.

fn generate_key<R: CryptoRng>(rng: &mut R) -> Key<Self>

👎Deprecated since 0.2.0:
use the Generate trait impl on Key instead

DEPRECATED: generate random key using the provided CryptoRng.

Instead, you can now use the Generate trait directly with the Key type:

let key = Key::generate_from_rng(rng);

fn generate_iv<R: CryptoRng>(rng: &mut R) -> Iv<Self>

👎Deprecated since 0.2.0:
use the Generate trait impl on Iv instead

DEPRECATED: generate random IV using the provided CryptoRng.

Instead, you can now use the Generate trait directly with the Iv type:

let iv = Iv::generate_from_rng(rng);

fn generate_key_iv<R: CryptoRng>(rng: &mut R) -> (Key<Self>, Iv<Self>)

👎Deprecated since 0.2.0:
use the Generate trait impls on Key and Iv instead

DEPRECATED: generate random key and IV using the provided CryptoRng.

Instead, you can now use the Generate trait directly with the Key and Iv types:

let key = Key::generate_from_rng(rng);
let iv = Iv::generate_from_rng(rng);

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.