Trait crypto_common::KeyIvInit

pub trait KeyIvInit: KeySizeUser + IvSizeUser + Sized {
    // Required method
    fn new(key: &Key<Self>, iv: &Iv<Self>) -> Self;

    // Provided method
    fn new_from_slices(key: &[u8], iv: &[u8]) -> Result<Self, InvalidLength> { ... }
}

Types which can be initialized from key and initialization vector (nonce).

Required Methods

fn new(key: &Key<Self>, iv: &Iv<Self>) -> Self

Create new value from fixed length key and nonce.

Provided Methods

fn new_from_slices(key: &[u8], iv: &[u8]) -> Result<Self, InvalidLength>

Create new value from variable length key and nonce.