Trait crypto_common::Serial

pub trait Serial {
    fn serial<B: Buffer>(&self, _out: &mut B);
}

Trait implemented by types which can be encoded into byte arrays. The intention is that the encoding is binary and not human readable.

Required methods

fn serial<B: Buffer>(&self, _out: &mut B)