Module crypto_common::types

Common types needed in concordium.

Structs

Amount
An amount of CCD. The lowest expressible amount is 1microCCD. The string representation of this type uses a decimal separator with at most 6 decimals.

CredentialIndex
Index of the credential that is to be used.

KeyIndex
Index of an account key that is to be used.

KeyPair
A ed25519 keypair. This is available in the ed25519::dalek crate, but the JSON serialization there is not compatible with what we use, so we redefine it there.

Memo

Signature
A single signature. Using the same binary and JSON serialization as the Haskell counterpart. In particular this means encoding the length as 2 bytes, and thus the largest size is 65535 bytes.

Timestamp
Datatype used to indicate a timestamp in milliseconds.

TransactionSignature
Transaction signature structure, to match the one on the Haskell side.

TransactionTime
Datatype used to indicate transaction expiry.

UrlText

Enums

AmountParseError
Errors that can occur during parsing of an Amount from a string.

DelegationTarget

OpenStatus

Constants

MAX_MEMO_SIZE

MAX_URL_SIZE